majorly restructured files to match best practices for Flask app
