+++ app/models/StatementOfActivities.py
@@ -1,10 +1,10 @@
-from utils.Table import Table
-from utils.Cell import Cell
-from utils.Context import Context
-from utils.Dimension import Dimension
-from utils.helper_functions import *
+from app.models.Table import Table
+from app.models.Cell import Cell
+from app.models.Context import Context
+from app.models.Dimension import Dimension
+from app.utils.helper_functions import *
 from typing import *
-from utils.constants import *
+from app.utils.constants import *
 
 class StatementofActivities(Table):
     """ extends the Sheet class for statement of activities """
