+++ app/models/Table.py
@@ -1,8 +1,8 @@
-from utils.Context import Context
-from utils.Cell import Cell
-from utils.helper_functions import *
+from app.models.Context import Context
+from app.models.Cell import Cell
+from app.utils.helper_functions import *
 import pandas as pd
-from utils.constants import *
+from app.utils.constants import *
 from typing import * # to specify funtion inputs and outputs
 from datetime import datetime # for date parsing
