+++ app/static/input_files/word_documents/CA Clayton 2022 Notes.docx
