+++ app/static/js/ixbrlviewer.js
