+++ app/static/js/spinner.js
