+++ app/static/output/ixbrl-viewer.html
