+++ app/templates/xbrl/base.html
