+++ app/templates/xbrl/table.html
