+++ app/templates/xbrl/word_contexts.html
