+++ app/templates/xbrl/xbrl_style.html
